+++ selfdrive/ui/tests/.gitignore
@@ -1,5 +1,6 @@
 test
 playsound
+snapshots
 test_sound
 test_translations
-test_ui_snapshot
+ui_snapshot
